requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary>=2.9
Werkzeug==2.3.7
python-dotenv==1.0.0
gunicorn>=21.2.0
requests
Flask-Mail==0.10.0
Flask-WTF
locust
flask-socketio
gevent
gevent-websocket
Flask-Limiter==3.5.0
